#1: slice mods are the new serialization in Go
